Read a Recipe in .xml Format

Read an .xml file that is read by the Minstrel liquid handling robot. Used to tell the robot what recipe to conduct. Individual components of the actual pipetting are calculated by the robot once recipe is given.

read_recipe(x)

Arguments

x

File

Value

a list with a tibble for $pipetting and a tibble for recipe components.

Examples

library(minstrel)

fl <- system.file("extdata", "test_plate.xml", package = "minstrel")

read_recipe(fl)
#> New names:
#> • `name` -> `name...2`
#> • `name` -> `name...10`
#> $pipetting
#> # A tibble: 120 × 12
#>    names name...2       pH    conc  cunits viscosity volatility barcode comments
#>    <chr> <chr>          <chr> <chr> <chr>  <chr>     <chr>      <chr>   <chr>   
#>  1 well  Bis-Tris prop… 6.2   1     M      5         50         ""      ""      
#>  2 well  Bis-Tris prop… 6.2   1     M      5         50         ""      ""      
#>  3 well  Bis-Tris prop… 6.2   1     M      5         50         ""      ""      
#>  4 well  Bis-Tris prop… 6.2   1     M      5         50         ""      ""      
#>  5 well  Bis-Tris prop… 6.2   1     M      5         50         ""      ""      
#>  6 well  Bis-Tris prop… 6.2   1     M      5         50         ""      ""      
#>  7 well  Bis-Tris prop… 6.2   1     M      5         50         ""      ""      
#>  8 well  Bis-Tris prop… 6.2   1     M      5         50         ""      ""      
#>  9 well  Bis-Tris prop… 6.2   1     M      5         50         ""      ""      
#> 10 well  Bis-Tris prop… 6.2   1     M      5         50         ""      ""      
#> # … with 110 more rows, and 3 more variables: name...10 <chr>, volume <chr>,
#> #   vunits <chr>
#> 
#> $stocks
#> # A tibble: 5 × 11
#>   name    pH    volume vunits conc  cunits viscosity volatility barcode comments
#>   <chr>   <chr> <chr>  <chr>  <chr> <chr>  <chr>     <chr>      <chr>   <chr>   
#> 1 Bis-Tr… "6.2" 1.49   ml     1     M      "5"       "50"       ""      ""      
#> 2 Bis-Tr… "9"   910.40 ul     1     M      "5"       "50"       ""      ""      
#> 3 magnes… ""    4.80   ml     1     M      "3"       "50"       ""      ""      
#> 4 polyet… ""    6.00   ml     100   v/v    "8"       "50"       ""      ""      
#> 5 water   ""    10.80  ml     100   v/v    ""        ""         ""      ""      
#> # … with 1 more variable: count <chr>
#>